Extrait un composant
