+++ app.territoiresentransitions.react/src/ui/shared/preuves/AddPreuveModal/AddFile.tsx
@@ -2,7 +2,7 @@
  * Affiche le composant d'upload de fichiers
  */
 import {ChangeEvent, FormEvent, useState} from 'react';
-import { Button, Checkbox, Field, InfoTooltip, Input } from '@tet/ui';
+import { Button, Field, Input } from '@tet/ui';
 import { HINT, EXPECTED_FORMATS_LIST } from './constants';
 import { filesToUploadList } from './filesToUploadList';
 import { TFileItem } from './FileItem';
@@ -14,6 +14,7 @@ import {
   UploadStatusCompleted,
 } from './types';
 import { useCollectiviteId } from 'core-logic/hooks/params';
+import { CheckboxConfidentiel } from './CheckboxConfidentiel';
 
 export type TAddFileFromLib = (fichier_id: number) => void;
 
@@ -25,9 +26,6 @@ export type TAddFileProps = {
   onAddFileFromLib: TAddFileFromLib;
   onClose: () => void;
 };
-
-// types de documents pour lesquels l'utilisateur peut choisir l'option "confidentiel"
-const ALLOW_PRIVATE: DocType[] = ['reglementaire', 'complementaire', 'annexe'];
 
 const getFileByName = (fileName: string, selection: Array<TFileItem>): number =>
   selection.findIndex(({ file }) => file.name === fileName);
@@ -100,27 +98,11 @@ export const AddFile = (props: TAddFileProps) => {
           onChange={onChange}
         />
       </Field>
-      {!!docType && ALLOW_PRIVATE.includes(docType) && (
-        <div className="flex flex-row items-center gap-2">
-          <Checkbox
-            variant="switch"
-            label="Document en mode privé"
-            checked={confidentiel}
-            onChange={(evt) => setConfidentiel(evt.currentTarget.checked)}
-          />
-          <InfoTooltip
-            iconClassName="text-primary-8"
-            className="whitespace-break-spaces !text-lg"
-            label={
-              docType === 'annexe'
-                ? MSG_ANNEXE_CONFIDENTIELLE
-                : MSG_DOC_CONFIDENTIEL
-            }
-            size="md"
-          />
-        </div>
-      )}
-
+      <CheckboxConfidentiel
+        docType={docType}
+        confidentiel={confidentiel}
+        setConfidentiel={setConfidentiel}
+      />
       <FileItemsList
         items={currentSelection}
         onRunningStopped={onRunningStopped}
@@ -138,14 +120,3 @@ export const AddFile = (props: TAddFileProps) => {
     </div>
   );
 };
-
-/** Libellés affichés dans l'infobulle à côté du bouton "confidentiel" */
-const MSG_ANNEXE_CONFIDENTIELLE = `Nous vous encourageons à partager vos documents : ils permettent à d’autres collectivités de s’inspirer de vos actions, de vos pratiques.
-
-Si vos documents sont confidentiels, vous pouvez activer cette option : seuls les membres de votre collectivité (dont votre conseiller et votre auditeur si vous êtes engagés dans le programme “Territoire Engagé Transition Écologique”) et le service support de la plateforme pourront y accéder.
-
-Si la fiche action est en mode privé, les documents ne seront pas accessibles par des personnes n’étant pas membres de votre collectivité, que le document soit en mode privé ou non.`;
-
-const MSG_DOC_CONFIDENTIEL = `Nous vous encourageons à partager vos documents : ils permettent à d’autres collectivités de s’inspirer de vos actions, de vos pratiques.
-
-Si vos documents sont confidentiels, vous pouvez activer cette option : seuls les membres de votre collectivité, votre conseiller, votre auditeur, les membres de la CNL (Commission National du Label) et le service support de la plateforme pourront y accéder.`;
